Move reduction_identity into its own header file
