+++ core/src/Kokkos_Complex.hpp
@@ -51,6 +51,7 @@
 #include <Kokkos_Atomic.hpp>
 #include <Kokkos_MathematicalFunctions.hpp>
 #include <Kokkos_NumericTraits.hpp>
+#include <Kokkos_ReductionIdentity.hpp>
 #include <impl/Kokkos_Error.hpp>
 #include <complex>
 #include <type_traits>
